NetNet2.0: Re-org net interface and message interface (DONE)

+++ http/nodeinfo/ngbnodeinfo.go
@@ -21,11 +21,11 @@ package nodeinfo
 import "strings"
 
 type NgbNodeInfo struct {
-	NgbId         string  //neighbor node id
+	NgbId         string //neighbor node id
 	NgbType       string
 	NgbAddr       string
 	HttpInfoAddr  string
-	HttpInfoPort  int
+	HttpInfoPort  uint16
 	HttpInfoStart bool
 }
